+++ packages/components/src/index.ts
@@ -95,6 +95,28 @@ export { default as LinkButton } from './molecules/LinkButton'
 export type { LinkButtonProps } from './molecules/LinkButton'
 export { default as Modal, ModalHeader, ModalBody } from './molecules/Modal'
 export type { ModalProps, ModalHeaderProps } from './molecules/Modal'
+export {
+  default as NavbarLinks,
+  NavbarLinksList,
+  NavbarLinksListItem,
+} from './molecules/NavbarLinks'
+export type {
+  NavbarLinksProps,
+  NavbarLinksListProps,
+  NavbarLinksListItemProps,
+} from './molecules/NavbarLinks'
+export {
+  default as NavbarSlider,
+  NavbarSliderHeader,
+  NavbarSliderContent,
+  NavbarSliderFooter,
+} from './molecules/NavbarSlider'
+export type {
+  NavbarSliderProps,
+  NavbarSliderHeaderProps,
+  NavbarSliderContentProps,
+  NavbarSliderFooterProps,
+} from './molecules/NavbarSlider'
 export { default as OrderSummary } from './molecules/OrderSummary'
 export type { OrderSummaryProps } from './molecules/OrderSummary'
 export {
@@ -235,6 +257,19 @@ export type {
   ImageGallerySelectorProps,
   ImageZoomProps,
 } from './organisms/ImageGallery'
+
+export {
+  default as Navbar,
+  NavbarHeader,
+  NavbarRow,
+  NavbarButtons,
+} from './organisms/Navbar'
+export type {
+  NavbarProps,
+  NavbarHeaderProps,
+  NavbarRowProps,
+  NavbarButtonsProps,
+} from './organisms/Navbar'
 
 export { default as OutOfStock } from './organisms/OutOfStock'
 export type { OutOfStockProps } from './organisms/OutOfStock'