+++ packages/core/src/components/common/Alert/Alert.tsx
@@ -1,17 +1,21 @@
 import type { PropsWithChildren, ReactNode } from 'react'
 import { useCallback, useState } from 'react'
 
-import { Alert as UIAlert, AlertProps } from '@faststore/ui'
+import { Alert as UIAlert, AlertProps as UIAlertProps } from '@faststore/ui'
 import { mark } from 'src/sdk/tests/mark'
 
-interface Props extends AlertProps {
+export interface AlertProps extends UIAlertProps {
   /**
    * For CMS integration purposes, should be used to pass content through it
    * instead pass through children
    */
   content?: ReactNode
 }
-function Alert(args: PropsWithChildren<Props>) {
+function Alert({
+  content,
+  children,
+  ...otherProps
+}: PropsWithChildren<AlertProps>) {
   const [displayAlert, setDisplayAlert] = useState(true)
 
   const onAlertClose = useCallback(
@@ -22,8 +26,6 @@ function Alert(args: PropsWithChildren<Props>) {
   if (displayAlert === false) {
     return null
   }
-
-  const { content, children, ...otherProps } = args
 
   return (
     <UIAlert onClose={onAlertClose} {...otherProps}>